Refactor source code structure.
