+++ appchain-anchor/src/assets/mod.rs
@@ -1,0 +1,2 @@
+pub mod near_fungible_tokens;
+mod wrapped_appchain_token;
