+++ appchain-anchor/src/assets/near_fungible_tokens.rs
