+++ appchain-anchor/src/assets/wrapped_appchain_token.rs
